Funcionalidad panel admin y nueva tabla para controlar acciones admins

+++ Back/database/migrations/2026_02_04_111419_update_levels_tables_add_test_cases.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -13,14 +12,22 @@ return new class extends Migration
     {
         // Actualizar tabla niveles_historia
         Schema::table('niveles_historia', function (Blueprint $table) {
-            $table->json('test_cases')->nullable()->after('descripcion'); 
-            $table->dropColumn('solucion_esperada');
+            if (!Schema::hasColumn('niveles_historia', 'test_cases')) {
+                $table->json('test_cases')->nullable()->after('descripcion');
+            }
+            if (Schema::hasColumn('niveles_historia', 'solucion_esperada')) {
+                $table->dropColumn('solucion_esperada');
+            }
         });
 
         // Actualizar tabla niveles_roguelike
         Schema::table('niveles_roguelike', function (Blueprint $table) {
-            $table->json('test_cases')->nullable()->after('descripcion');
-            $table->dropColumn('codigo_validador');
+            if (!Schema::hasColumn('niveles_roguelike', 'test_cases')) {
+                $table->json('test_cases')->nullable()->after('descripcion');
+            }
+            if (Schema::hasColumn('niveles_roguelike', 'codigo_validador')) {
+                $table->dropColumn('codigo_validador');
+            }
         });
     }
